fix(yield): don't rely on env values
This removes the mechanism of putting the block in the env. To access
the block it is attached to self.

This approach is not very clean, as `self` doesn't need to have a block.
It's more like a workaround for now as it is the most safe way of
providing the block to method blocks. This will most likely change in
the future if we have a better abstraction of scope/environment/binding.